Change: Type of SourceLocation::line from std::size_t to std::uint64_t

+++ ltnc/SourceLocation.hxx
@@ -4,13 +4,13 @@
 
 namespace ltn::c {
 	struct SourceLocation {
-		std::size_t line = 0;
+		std::uint64_t line = 0;
 		std::string sourcename = "";
 	};
 
 
 	struct ArgumentLocation {
 		SourceLocation source_location;
-		std::size_t index;
+		std::uint64_t index;
 	};
 }
